fix settings swagger and remove useless
