+++ package.json
@@ -31,7 +31,6 @@
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/swagger": "^7.3.1",
     "@nestjs/throttler": "^5.1.2",
-    "@types/serve-static": "^1.15.7",
     "bcryptjs": "^2.4.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
@@ -40,8 +39,7 @@
     "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.2.0",
-    "rxjs": "^7.8.1",
-    "serve-static": "^1.15.0"
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.1.1",
