+++ src/app.module.ts
@@ -1,14 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { ThrottlerGuard } from '@nestjs/throttler';
 import { appModuleImports } from './app.module.imports';
 
 @Module({
   imports: appModuleImports,
-  controllers: [AppController],
   providers: [
-    AppService,
     {
       provide: 'APP_GUARD',
       useClass: ThrottlerGuard,
