+++ src/utils/templates/root-html-template.ts
@@ -1,0 +1,15 @@
+export const getHtmlTemplate = (host: string): string => `
+    <!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Welcome</title>
+    </head>
+    <body>
+    <h1>Hello!</h1>
+    <p>Swagger Shoplist API: <a href="http://${host}:3000/swagger">API</a></p>
+<p>Visit my GitHub project: <a href="https://github.com/s0vit/shoplist-be">Shoplist</a></p>
+</body>
+</html>
+    `;
