--- types/index.d.ts
@@ -1,13 +1,0 @@
-//index.d.ts
-import { Request } from 'express';
-
-declare global {
-  namespace Express {
-    export interface Request {
-      user: {
-        userId: string;
-        email: string;
-      };
-    }
-  }
-}
